FBO: Ubyte index texture

+++ src/glsl/mouse.frag.js
@@ -1,6 +1,6 @@
 export default /* glsl */`
 precision highp float;
-uniform int psel;
+uniform float psel;
 uniform vec2 tSize;
 uniform vec3 mouse;
 uniform sampler2D tPosition;
@@ -20,7 +20,7 @@ void main() {
 	vec2 uv = gl_FragCoord.xy / tSize.xy;
 	vec3 pos = texture2D( tPosition, uv ).xyz;
 	vec3 org = texture2D( tOriginal, uv ).xyz;
-	uv = getUV( float( psel ) );
+    uv = getUV( psel );
 	vec3 ref = texture2D( tOriginal, uv ).xyz;
 	vec3 offset = mouse - ref;
 	if ( distance( org, ref ) <= 05.0 )  {
